Update patch set 2
Patch Set 2: Code-Review+1

Interesting idea.  I originally accomplished this using a somewhat-obnoxious-yet-supported syntax in the config file:

    motd = This is a very long MOTD! \n\
It can occupy several lines!

My only question is if the config lines are always guaranteed to be returned in the order in which they are entered in the config.  If they are, then I agree, this is a MUCH cleaner method to the hack above!

Patch-set: 2