Update patch set 2
Patch Set 2:

Because if it's been added as a custom plugin, the parent repository ends up being dirty (uncommitted changes in a submodule). The other option is putting it in .gitignore, but the file looked generic enough that I thought it made more sense to commit it.

Patch-set: 2
Reviewer: Chad Horohoe <1010813@173816e5-2b9a-37c3-8a2e-48639d4f1153>